backend/requirements.txt
# Django and DRF
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0

# Database
psycopg2-binary==2.9.9

# Environment management
python-decouple==3.8

# Celery and Redis
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1

# PDF Generation
reportlab==4.0.7
weasyprint==60.2

# Email
django-anymail==10.2

# Development and Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# API Documentation
drf-yasg==1.21.7

# File handling
Pillow==10.1.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
